coderoom-backend/requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
psycopg2-binary
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
